+++ NEWS.md
@@ -1,3 +1,8 @@
+## version 3.2.1
+### Features
+- default prior distributions for all parameters are now set via the `set_default_priors()` function
+- `rescale_priors` argument allows to conveniently re-scale the prior distributions for the effect, heterogeneity, and bias simultaneously
+
 ## version 3.2
 ### Features
 - `summary_heterogeneity()` function to summarize the heterogeneity of the RoBMA models (prediction interval, tau, tau^2, I^2, and H^2)